Added platform specific bytesToTryteValues
